requirements.txt
pyspark==3.5.0
cryptography==43.0.3
delta-spark==3.2.0
boto3==1.35.18